extract solana-signature crate
