+++ sdk/Cargo.toml
@@ -23,7 +23,6 @@ default = [
 full = [
     "byteorder",
     "chrono",
-    "generic-array",
     "memmap2",
     "rand",
     "rand0-7",
@@ -91,6 +90,7 @@ solana-frozen-abi-macro = { workspace = true, optional = true }
 solana-program = { workspace = true }
 solana-program-memory = { workspace = true }
 solana-sanitize = { workspace = true }
+solana-signature = { workspace = true }
 solana-sdk-macro = { workspace = true }
 solana-secp256k1-recover = { workspace = true }
 solana-serde-varint = { workspace = true }
