+++ sdk/signature/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "solana-signature"
+description = "Solana 64-byte signature type"
+documentation = "https://docs.rs/solana-signature"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+bs58 = { workspace = true }
+ed25519-dalek = { workspace = true }
+generic-array = { workspace = true, features = ["serde", "more_lengths"] }
+rand = { workspace = true }
+serde = { workspace = true }
+serde_derive = { workspace = true }
+solana-frozen-abi = { workspace = true, optional = true }
+solana-frozen-abi-macro = { workspace = true, optional = true }
+solana-sanitize = { workspace = true }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+curve25519-dalek = { workspace = true }
+solana-program = { workspace = true }
+solana-sdk = { workspace = true }
+
+[build-dependencies]
+rustc_version = { workspace = true }
+
+[features]
+frozen-abi = ["dep:solana-frozen-abi", "dep:solana-frozen-abi-macro"]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
